requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
